Create a journal page for blogging and index for homepage
